+++ src/pages/PostedPage/components/MessageCardList/MessageCardList.jsx
@@ -2,8 +2,17 @@ import { Link } from 'react-router-dom';
 import Button from '@/components/Button/Button';
 import MessageCard from '@/pages/PostedPage/components/MessageCard/MessageCard';
 import { MOCK_MESSAGES } from '@/pages/PostedPage/components/MessageCardList/messages';
+import Skeleton from '@/components/Skeleton/Skeleton';
+import useInfiniteScrollMessages from '@/pages/PostedPage/hooks/useInfiniteScroll';
 
 const MessageCardList = () => {
+  const { displayedMessages, observerRef, isLoading } =
+    useInfiniteScrollMessages({
+      data: MOCK_MESSAGES,
+      initialCount: 5,
+      nextCount: 3,
+    });
+
   return (
     <div className="grid grid-cols-1 sm:grid-cols-2 xl:grid-cols-3 gap-6">
       {/* 첫 번째 메세지 카드 고정*/}
@@ -15,7 +24,7 @@ const MessageCardList = () => {
         <Button shape="circle" iconName="plus" variant="gray" />
       </Link>
       {/* 카드 리스트 */}
-      {MOCK_MESSAGES.map((msg) => (
+      {displayedMessages.map((msg) => (
         <MessageCard
           key={msg.id}
           sender={msg.sender}
@@ -25,6 +34,24 @@ const MessageCardList = () => {
           createdAt={msg.createdAt}
         />
       ))}
+
+      {/* 감시용 div (맨 아래) */}
+      <div ref={observerRef} className="h-[1px] col-span-full bg-transparent" />
+
+      {/* 로딩  스켈레톤 */}
+      {isLoading && (
+        <>
+          {Array.from({ length: 3 }).map((_z, i) => (
+            <Skeleton
+              key={i}
+              variant="base"
+              shape="rect"
+              width="100%"
+              height="280px"
+            />
+          ))}
+        </>
+      )}
     </div>
   );
 };
